sks ta: utils to sanitize a client serialized object
When a client provides an object description to the SKS, it
provides a buffer storing attributes in a serialized format
as defined by the TA API. This change provides function to
generate a sanitized serial object out of the client proposed
object (called template in pkcs11 literature).

The client serialized object is a memory buffer starting with
a header structure followed by each attribute content (id, size
and value).

Add a debug function to dump the content of a client serialized
object.

Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>